Permit variation mounts in mounts.json
First, I changed the schema to allow `variation` and removed `description`
as a required field (changing `additionalProperties` to `true` as a result).

Secondly, every mount has a `variation`, set to `true` or `false` depending
on if it is the base skin or not.  For mounts that have a `Model` name that
matches an entry in `GameStrings.txt`, I use that for the `description`.

There are three (3) different schemas of variations for mounts within the
XML (thanks blizz!).  Two (2) of them reference an original base model and
the third reference a common model. Only the third method populates all the
fields as we cannot reference other entries within the `.map()` loop.  To
completely fill these, a second iteration over the mounts will be needed.